Tox + Pylint cleanup
Fixes a bug which would result in black and pylint disagreeing forever.
(owing to pylints bad continuation warning, see #20).

Prevents pylint from failing if there are `# TODO: ...` comments in the
code, because I (personally) think there is nothing wrong with
documenting tech debt or future plans even if you can't deal with them
before a commit/merge.

Adds a command to explicitly list/check for `TODO` comments.

Fixes the command to run isort being the same as the command to check if
isort needed to be run (copy paste oops :( )

- Closes #21
- Closes #20
- Closes #19