Dev Environment Setup
Provides copy/pastable commands for setup after the template is created
in addition to copy/pastable commands in the generated projects for dev
environment creation.

Refactors local tox to run against all supported python versions (at the
moment that is 36, 37, 38)

Moves to requiring xxenv-latest, rather than pyenv-install-latest as
part of the setup instructions.

Fixes a deprecated CLI arg in the `tasks.py` which would emit a warning.