Move requirements to setup.py
Moves all requirements to the `extras_require` section of the `setup.py`
in order to avoid having several different requirements files (and the
`requirements/` directory in the top level of the repo).

Requirements from `requirements/ci.txt` have been moved into the CI
workflow itself, rather than adding an `extras_require` entry for it
(because they should never need to be installed anywhere _but_ CI).

This has the effect of slightly changing the documentation regarding
project setup + the way tox installs these requirements.

Adds some default configurations to the `tasks.py` to keep the output of
commands pretty + print the commands that are being run.

Fixes a bug with the `inv.pindeps` command where if a projects pip name
differed from its module name the editable installation would be
included in `requirements.txt`