Docs Update
Updates the docs of the generated projects.

- The docs now use the [Diátaxis Documentation Framework](https://diataxis.fr/).
- Moved the autodocs into the reference section of the new docs layout
- Add `sphinx.ext.todo` and `sphinx_autodoc_typehints` to the sphinx
  extensions, with relevant configuration.
- Significantly updates the `inv serve.docs` command
  - It now optionally renders the `todos` in the dev docs
  - It now has a (repeatable) flag for watching additional directories
  - It now has a flag for automatically opening the browser

Adds an editorconfig configuration for `rst` files to set indents to 3
spaces.

Adds a few more checkboxes to the PR template, specifically one for
tests passing locally and two for updating the docs.

Updates the project's README to include some more relevant information
and links for recent template changes.

Updates (and significantly expands) the `.gitignore` in the generated
project, utilizing gitignore.io.