Remove setup.py + Remove optional docs flag
Removes `setup.py`, as after pip version 21.1 it is possible to editable
install packages without one if you have a `setup.cfg` and a
`pyproject.toml`!

Replaces all data in `setup.py` with a declarative `setup.cfg`.

Additionally, removes the flag for the docs folder. The docs folder has
been pretty tightly integrated into multiple places in the template, and
I almost always use a docs folder. As such, the effort of keeping it
properly parametrized in the template was becoming a bigger and bigger
lift. If there is some reason to parametrize the docs folder again in
the future I'll consider it.