Add pre-commit
Adds `pre-commit` to the template.

Swaps `tox` for `nox`

Adds `poetry`

Moves `mypy` config to `pyproject.toml`

Adds `py.typed` to the package, to carry forward type-hints.

Added typehints where required for strict checking to pass.

Cleans up the `inv clean` command.

These were a couple of big changes, primarily motivated by the need to
get proper locking in place and use the same versions of the
dependencies in the tests as are used during runtime.