Fix readthedocs after poetry switch
Fix readthedocs builds after switching to poetry.

Makes the requirements for docs and tests extras, instead of just dev
dependencies, so they can be installed by rtd.

There also appears to be a weird issue with `jinja2` and
`typing-extenions` in the latex build for the docs. For now I'm just
adding it manually to the documentation requirements.

Makes the required changes to the documentation to install them as
extras in appropriate circumstances.

Refactors the noxfile to take these changes into account.